perf: GPU scatter wave 1+2 optimizations
Wave 1 — CPU quick wins:
- Phase 1a: Grid spatial index (128×128) for O(1) hit testing (was O(n) linear scan)
- Phase 1b: Workgroup size 64 (was 256) — optimal for Apple Silicon M-series
- Phase 2: unorm8x4 color packing — 4 bytes/point vs 16 (4× bandwidth reduction)
  - CPU packs Uint32Array instead of Float32Array
  - WGSL unpackColor fn decodes in vertex shader
- Phase 8: Alpha-to-coverage AA skipped (needs MSAA render target restructure)

Wave 2 — Phase 3: uint8 RGBA wire format for continuous colors:
- Backend sends rgba uint8 (4 bytes/pt) instead of float32 (16 bytes/pt)
- Frontend parses as Uint8Array, zero-copy Uint32Array view for GPU upload
- 4× reduction in continuous color payload size

+++ frontend/src/scatter-gpu/components/ScatterGPUHost.tsx
@@ -17,7 +17,7 @@ import type { ScatterData, ScatterplotConfig, ScatterplotHandle } from "../types
 
 export interface ScatterGPUHostHandle {
     setColors(palette: readonly (readonly [number, number, number])[], indices?: Uint8Array): void;
-    setColorsDirect(rgba: Float32Array): void;
+    setColorsDirect(rgba: Uint8Array): void;
     getViewState(): { panX: number; panY: number; zoom: number };
     worldToScreen(wx: number, wy: number, w: number, h: number): { x: number; y: number };
 }
